fix(nutrition): round-2レビュー指摘(#1046)を解消
Fable5 round-1レビューのCritical/Warning/Suggestionを反映。

- [Critical] src/app/api/health/goals/[id]/route.tsからcalculateGoalProgress
  Percentageをvalue exportしていたため、Next.js 14.2のNextTypesPluginが
  route entry exportをHTTPメソッド+config系に限定する型チェックでnext build
  (=Vercelデプロイ)が失敗する不具合を修正。関数をsrc/lib/health-goal-progress.ts
  に切り出し、[id]/route.ts・route.ts(POST)・tests/health-goal-progress-
  direction.test.tsのimportを差し替え(ロジックは無変更)。`next build`を実際に
  実行し、型チェック段階(Linting and checking validity of types)を通過する
  ことを確認済み(以降のstatic export失敗はsandbox内でSupabase env未設定の
  ためで本修正とは無関係)。
- [Warning] regenerate-shopping-list-v2のbuildPromptが、集計後の
  InputIngredient(name/amount/count)をそのままLLMに渡しており、amountが
  「確定合算済みの総量」に変わった後もcountの意味を定義していなかったため
  LLMがamount×countと誤読し購入量を水増しする恐れがあった。LLM入力から
  countを除去しname/amountのみを渡すよう変更し、プロンプトにも
  「amountは既に合算済みの確定総量」である旨を明記。
- [Suggestion] goals/route.ts(POST)の進捗率計算の死んだ分岐を、PUT側と
  同じcalculateGoalProgressPercentageに置き換えて重複ロジックを解消
  (作成時はcurrent_value=start_valueのため結果は従来通り常に0%)。

以下はround-1レビューの通り対応不要と判断(反証・報告のみ):
- [Warning] health-payloads.tsのmood_score等1-5スケール化は意図的な
  fail-closedで、既存6-10データ移行はmigration変更禁止のスコープ外
  (spec: wave6-common.md)。DB移行方針決定・CHECK制約追加のfollow-up Issue
  起票が別途必要(オーケストレーター側で起票要)。
- [Suggestion] apps/mobile/src/components/menu/RadarChart.tsxの平均達成率
  にもUX3-20と同型の上限系栄養素混入が残るが、apps/mobileは変更禁止
  (spec: wave6-common.md、#1079競合)のため報告のみ。本コミットメッセージに
  明記することでreport-only項目として記録する。
- [Suggestion] nutrition-calculator.ts(v1)のF5-11修正はcalculateNutrition
  FromIngredientsがどこからも呼ばれていない未使用コードであることをgrepで
  再確認済み(round-1同様、本番挙動への影響なし)。

テスト: npx tsc --noEmit / eslint 対象ファイルでエラーなし。
vitest run 全112ファイル998テストPASS(既存回帰なし)。
